feat(server): add Prometheus metrics endpoint
This commit introduces a `/metrics` endpoint for Prometheus to scrape stream
metrics, including FPS and average FPS per stream. Additionally, it adds the
`--stream-id-label` argument, allowing users to optionally include the `stream-id`
label in Prometheus metrics.
