+++ server/utils/__init__.py
@@ -1,0 +1,2 @@
+from .utils import patch_loop_datagram, add_prefix_to_app_routes
+from .fps_meter import FPSMeter
